+++ src/main/java/qengine/program/teamengine/index/pso/PSO.java
@@ -1,19 +1,19 @@
-package qengine.program.teamengine.index.OPS;
+package qengine.program.teamengine.index.pso;
 
 import qengine.program.teamengine.index.MyIndex;
 
 import java.util.List;
 
-public class OPS extends MyIndex {
-    private static OPS instance = null;
+public class PSO extends MyIndex {
+    private static PSO instance = null;
 
-    private OPS() {
+    private PSO() {
         super();
     }
 
-    public static OPS getInstance() {
+    public static PSO getInstance() {
         if (instance == null) {
-            instance = new OPS();
+            instance = new PSO();
         }
         return instance;
     }
@@ -23,22 +23,22 @@ public class OPS extends MyIndex {
         int subjectIndex = dictionary.getIndexFromElement(subject);
         int predicateIndex = dictionary.getIndexFromElement(predicate);
         int objectIndex = dictionary.getIndexFromElement(object);
-        this.addStatementToIndex(objectIndex, predicateIndex, subjectIndex);
+        this.addStatementToIndex(predicateIndex, subjectIndex, objectIndex);
         TIMERS.addTimerToIndexes();
     }
 
     @Override
     public List<Integer> getRes(int subject, int predicate, int object) {
-        return getResGeneral(object, predicate);
+        return getResGeneral(predicate, subject);
     }
 
     @Override
     public List<Integer> secondRes(int subject, int predicate, int object, List<Integer> oldRes) {
-        return getSecondResGeneral(object, predicate, oldRes);
+        return getSecondResGeneral(predicate, subject, oldRes);
     }
 
     @Override
     public String toString() {
-        return "OPS";
+        return "PSO";
     }
 }
