+++ src/main/java/qengine/program/teamengine/process/ProcessQuery.java
@@ -3,13 +3,13 @@ package qengine.program.teamengine.process;
 import org.eclipse.rdf4j.query.algebra.StatementPattern;
 import org.eclipse.rdf4j.query.algebra.Var;
 import qengine.program.teamengine.index.MyIndex;
-import qengine.program.teamengine.index.OPS.OPS;
-import qengine.program.teamengine.index.OSP.OSP;
-import qengine.program.teamengine.index.POS.POS;
-import qengine.program.teamengine.index.PSO.PSO;
-import qengine.program.teamengine.index.SOP.SOP;
-import qengine.program.teamengine.index.SPO.SPO;
-import qengine.program.teamengine.q1.Dictionary;
+import qengine.program.teamengine.index.ops.OPS;
+import qengine.program.teamengine.index.osp.OSP;
+import qengine.program.teamengine.index.pos.POS;
+import qengine.program.teamengine.index.pso.PSO;
+import qengine.program.teamengine.index.sop.SOP;
+import qengine.program.teamengine.index.spo.SPO;
+import qengine.program.teamengine.dictionary.Dictionary;
 
 import java.util.List;
 import java.util.stream.Collectors;
